pyproject.toml
[project]
name = "company-lens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.1",
    "openai>=2.36.0",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.1",
    "python-dotenv>=1.2.2",
    "scalar-fastapi>=1.8.2",
    "sqlmodel>=0.0.38",
    "tavily-python>=0.7.24",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "SIM"]
